refactor: Member 연관 관계 제거 후 발생되는 테스트 에러 조치
