+++ src/test/java/kr/pickple/back/member/service/MemberServiceTest.java
@@ -1,11 +1,8 @@
 package kr.pickple.back.member.service;
 
-import static kr.pickple.back.common.domain.RegistrationStatus.*;
-import static kr.pickple.back.member.exception.MemberExceptionCode.*;
 import static org.assertj.core.api.Assertions.*;
 import static org.mockito.BDDMockito.*;
 
-import java.util.List;
 import java.util.Optional;
 
 import org.junit.jupiter.api.DisplayName;
@@ -25,14 +22,12 @@ import kr.pickple.back.auth.domain.token.AuthTokens;
 import kr.pickple.back.auth.domain.token.JwtProvider;
 import kr.pickple.back.auth.domain.token.RefreshToken;
 import kr.pickple.back.auth.repository.RedisRepository;
-import kr.pickple.back.crew.dto.response.CrewProfileResponse;
 import kr.pickple.back.fixture.domain.MemberFixtures;
 import kr.pickple.back.fixture.dto.MemberDtoFixtures;
 import kr.pickple.back.member.domain.Member;
 import kr.pickple.back.member.dto.request.MemberCreateRequest;
 import kr.pickple.back.member.dto.response.AuthenticatedMemberResponse;
 import kr.pickple.back.member.dto.response.MemberProfileResponse;
-import kr.pickple.back.member.exception.MemberException;
 import kr.pickple.back.member.repository.MemberRepository;
 
 @ExtendWith(MockitoExtension.class)
@@ -106,59 +101,6 @@ class MemberServiceTest {
 
         // then
         assertThat(memberProfileResponse).isNotNull();
-    }
-
-    @Test
-    @DisplayName("회원이 가입한 크루 목록을 조회할 수 있다.")
-    void findAllCrewsByMemberId_ReturnCrewProfileResponses() {
-        // given
-        final Long memberId = 1L;
-        final Long loggedInMemberId = 1L;
-        final Member member = buildMember();
-        given(memberRepository.findById(anyLong())).willReturn(Optional.ofNullable(member));
-
-        // when
-        final List<CrewProfileResponse> crewProfileResponses = memberService.findAllCrewsByMemberId(memberId,
-                loggedInMemberId, CONFIRMED);
-
-        // then
-        assertThat(crewProfileResponses).isNotNull();
-    }
-
-    @Test
-    @DisplayName("회원이 만든 크루 목록을 조회할 수 있다.")
-    void findCreatedCrewsByMemberId_ReturnCrewProfileResponses() {
-        // given
-        final Long memberId = 1L;
-        final Long loggedInMemberId = 1L;
-        final Member member = buildMember();
-
-        given(memberRepository.findById(anyLong())).willReturn(Optional.ofNullable(member));
-
-        // when
-        final List<CrewProfileResponse> crewProfileResponses = memberService.findCreatedCrewsByMemberId(
-                loggedInMemberId,
-                memberId
-        );
-
-        // then
-        assertThat(crewProfileResponses).isNotNull();
-    }
-
-    @Test
-    @DisplayName("회원이 만든 크루 목록을 조회할 때 본인이 만든 크루가 아닌 경우 예외가 발생한다.")
-    void findCreatedCrewsByMemberId_ThrowException() {
-        // given
-        final Long memberId = 1L;
-        final Long loggedInMemberId = 2L;
-        final Member member = buildMember();
-
-        // when && then
-        assertThatThrownBy(() -> memberService.findCreatedCrewsByMemberId(
-                loggedInMemberId,
-                memberId
-        )).isInstanceOf(MemberException.class)
-                .hasMessage(MEMBER_MISMATCH.getMessage());
     }
 
     private Member buildMember() {
